feat: support lido modules
Added new types in `crates/common/src/types.rs` to model the different
Lido modules we'll encounter depending on the Chain.

Also added a `HashMap` that introduces a new dimension to the pseudo-map
that previously always returned the registry address for the Lido
curated module in `crates/common/src/config/mux.rs`, with the goal of
being able to return the corresponding address based on the Chain and
Lido Module id.

To specify the required Lido module, a new optional field was added to
`examples/configs/pbs_mux.toml` for lido loader, called `lido_module_id`
which always defaults to 1.