feat: support lido community staking module
Added support for Lido's Community Staking Module (CSM), since the
deployed contract has a different ABI, so the same registry used for the
other modules cannot be used.

Because this difference in ABIs, the CSM contract ABI has been imported,
containing the required functions. One key difference is that in the CSM
contract, `getTotalSigningKeyCount´ does not exist; therefore, for CSM
this call is replaced with `getNodeOperatorSummary´, which returns a
tuple of several values, including `totalDepositedValidators´ and
`depositableValidatorsCount´, whose sum is used to compute the total
number of keys.

Another difference between ABIs is the `getSigningKeys´ method, which in
CSM returns directly a `Bytes´ type.

The logic for obtaining `total_keys´ and the iteration to fetch the keys
has been extracted into smaller functions since this new case for CSM
encourages code duplication; therefore, the main function
`fetch_lido_registry_keys´ clearly separate both cases.